Logging refactor + formatting
