+++ apps/optimize-draft-listing-image/jest.config.js
@@ -1,10 +1,7 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  roots: ['<rootDir>/src'],
-  testMatch: ['**/__tests__/**/*.test.ts', '**/?(*.)+(spec|test).ts'],
-  collectCoverageFrom: [
-    'src/**/*.ts',
-    '!src/**/*.d.ts',
-  ],
+	preset: "ts-jest",
+	testEnvironment: "node",
+	roots: ["<rootDir>/src"],
+	testMatch: ["**/__tests__/**/*.test.ts", "**/?(*.)+(spec|test).ts"],
+	collectCoverageFrom: ["src/**/*.ts", "!src/**/*.d.ts"],
 };
