+++ apps/optimize-draft-listing-image/src/logger.ts
@@ -1,232 +1,46 @@
 /** biome-ignore-all lint/suspicious/noExplicitAny: <explanation> */
-import type { Context } from "aws-lambda";
-import winston from "winston";
+import type winston from "winston";
 
-export class Logger {
-	private winston: winston.Logger;
-	private context?: Context;
+export abstract class BaseLogger {
+	protected winston: winston.Logger;
 
-	constructor(context?: Context) {
-		this.context = context;
-		this.winston = winston.createLogger({
-			level: process.env.LOG_LEVEL || "debug",
-			format: winston.format.combine(
-				winston.format.timestamp(),
-				winston.format.errors({ stack: true }),
-				winston.format.json(),
-				winston.format.printf(({ timestamp, level, message, ...meta }) => {
-					const contextInfo = this.context
-						? {
-								requestId: this.context.awsRequestId,
-								functionName: this.context.functionName,
-								remainingTime: this.context.getRemainingTimeInMillis?.(),
-							}
-						: {};
-
-					return JSON.stringify({
-						timestamp,
-						level,
-						message,
-						context: contextInfo,
-						...meta,
-					});
-				}),
-			),
-			transports: [new winston.transports.Console()],
-		});
+	constructor(winston: winston.Logger) {
+		this.winston = winston;
 	}
 
-	info(message: string, ...meta: any[]): Logger;
-	info(message: any): Logger;
-	info(infoObject: object): Logger;
-	info(message: any, ...meta: any[]): Logger {
+	info(message: string, ...meta: any[]): BaseLogger;
+	info(message: any): BaseLogger;
+	info(infoObject: object): BaseLogger;
+	info(message: any, ...meta: any[]): BaseLogger {
 		this.winston.info(message, ...meta);
 		return this;
 	}
 
-	error(message: string, ...meta: any[]): Logger;
-	error(message: any): Logger;
-	error(infoObject: object): Logger;
-	error(message: any, ...meta: any[]): Logger {
+	error(message: string, ...meta: any[]): BaseLogger;
+	error(message: any): BaseLogger;
+	error(infoObject: object): BaseLogger;
+	error(message: any, ...meta: any[]): BaseLogger {
 		this.winston.error(message, ...meta);
 		return this;
 	}
 
-	warn(message: string, ...meta: any[]): Logger;
-	warn(message: any): Logger;
-	warn(infoObject: object): Logger;
-	warn(message: any, ...meta: any[]): Logger {
+	warn(message: string, ...meta: any[]): BaseLogger;
+	warn(message: any): BaseLogger;
+	warn(infoObject: object): BaseLogger;
+	warn(message: any, ...meta: any[]): BaseLogger {
 		this.winston.warn(message, ...meta);
 		return this;
 	}
 
-	debug(message: string, ...meta: any[]): Logger;
-	debug(message: any): Logger;
-	debug(infoObject: object): Logger;
-	debug(message: any, ...meta: any[]): Logger {
+	debug(message: string, ...meta: any[]): BaseLogger;
+	debug(message: any): BaseLogger;
+	debug(infoObject: object): BaseLogger;
+	debug(message: any, ...meta: any[]): BaseLogger {
 		this.winston.debug(message, ...meta);
 		return this;
 	}
 
-	logLambdaStart(): void {
-		if (!this.context) return;
-
-		this.info("Lambda Handler Started", {
-			functionVersion: this.context.functionVersion,
-			memoryLimitMB: this.context.memoryLimitInMB,
-		});
+	child(context: object): winston.Logger {
+		return this.winston.child(context);
 	}
-
-	logLambdaEnd(processedCount: number, errorCount: number): void {
-		this.info("Lambda Handler Completed", {
-			processedCount,
-			errorCount,
-		});
-	}
-
-	logRecordStart(index: number, total: number, messageId: string): void {
-		this.info("Processing SQS Record", {
-			recordIndex: index + 1,
-			totalRecords: total,
-			messageId,
-		});
-	}
-
-	logS3Event(
-		eventName: string,
-		bucketName: string,
-		objectKey: string,
-		objectSize?: number,
-	): void {
-		this.info("Processing S3 record", {
-			eventName,
-			bucketName,
-			objectKey,
-			objectSize,
-		});
-	}
-
-	logOptimizationStart(bucketName: string, objectKey: string): void {
-		this.info("Starting S3 image optimization", {
-			s3Location: `s3://${bucketName}/${objectKey}`,
-		});
-	}
-
-	logDownloadStart(bucketName: string, objectKey: string): void {
-		this.info("Downloading image from S3", {
-			s3Location: `s3://${bucketName}/${objectKey}`,
-		});
-	}
-
-	logDownloadComplete(sizeKB: string, duration: number): void {
-		this.info("Image downloaded successfully", {
-			sizeKB,
-			durationMs: duration,
-		});
-	}
-
-	logImageMetadata(metadata: any, duration: number): void {
-		this.info("Original image metadata analyzed", {
-			...metadata,
-			analysisDurationMs: duration,
-		});
-	}
-
-	logResponsiveSizes(
-		sizes: Array<{ name: string; width: number; height: number }>,
-	): void {
-		this.info("Generated responsive sizes", {
-			sizes: sizes.map((s) => ({
-				name: s.name,
-				dimensions: `${s.width}x${s.height}`,
-			})),
-		});
-	}
-
-	logProcessingStart(totalVariants: number): void {
-		this.info("Starting image variant optimization", {
-			totalVariants,
-			strategy: "serial compression with parallel uploads",
-		});
-	}
-
-	logProcessingComplete(): void {
-		this.info("All image variants processed successfully");
-	}
-
-	logTimingBreakdown(timings: {
-		download: number;
-		metadata: number;
-		processing: number;
-		total: number;
-	}): void {
-		this.info("Processing timing breakdown", {
-			downloadMs: timings.download,
-			metadataMs: timings.metadata,
-			processingMs: timings.processing,
-			totalMs: timings.total,
-		});
-	}
-
-	logVariantProcessing(
-		format: string,
-		sizeName: string,
-		width: number,
-		height: number,
-	): void {
-		this.info("Processing image variant", {
-			format,
-			sizeName,
-			dimensions: `${width}x${height}`,
-		});
-	}
-
-	logVariantComplete(
-		format: string,
-		sizeName: string,
-		sizeKB: string,
-		duration: number,
-	): void {
-		this.info("Image variant processed", {
-			format,
-			sizeName,
-			sizeKB,
-			durationMs: duration,
-		});
-	}
-
-	logUploadQueued(format: string, key: string): void {
-		this.info("Upload queued", {
-			format,
-			s3Key: key,
-		});
-	}
-
-	logUploadComplete(
-		format: string,
-		sizeName: string,
-		key: string,
-		uploadDuration: number,
-		totalDuration: number,
-	): void {
-		this.info("Upload completed", {
-			format,
-			sizeName,
-			s3Key: key,
-			uploadDurationMs: uploadDuration,
-			totalDurationMs: totalDuration,
-		});
-	}
-
-	logUploadSkipped(format: string, key: string, sizeKB: string): void {
-		this.info("Upload skipped - UPLOAD_TO_S3 disabled", {
-			format,
-			s3Key: key,
-			sizeKB,
-		});
-	}
-}
-
-export function bytesToKB(bytes: number): string {
-	return (bytes / 1024).toFixed(2);
 }
